Move clear-win top-level extensions into companion objects
Relocate top-level extension methods into the companion objects of
types appearing in their receiver signatures, so they're discovered
through implicit scope without needing to be re-exported from
`soundness_*.scala`.

- `Dag[Text].dot` → `acyclicity.Dag`
- `Diff[Text].serialize` → `dissonance.Diff`
- `Encoding{type CanEncode = true}.encoder` → `hieroglyph.Encoding`
- `Seq[Html].nodes` → `honeycomb.Html`
- `Seq[Task[r]].sequence`, `Iterable[Task[r]].race` → `parasite.Task`
- `(Markdown of Layout).sections` → `punctuation.Markdown`
- `Quantity[u].{in, invert, normalize, sqrt, cbrt, units, express, dimension}` → `quantitative.internal.Quantity`
- `Cookie[Session].session` → `telekinesis.Cookie`
- `Seq[Xml].nodes` → `xylophone.Xml`
