+++ lib/acyclicity/src/core/acyclicity_core.scala
@@ -32,14 +32,10 @@
                                                                                                   */
 package acyclicity
 
-import anticipation.*
 import denominative.*
 import proscenium.*
 import rudiments.*
 import spectacular.*
-
-extension (dag: Dag[Text])
-  def dot: Dot = Digraph(dag.edges.to(List).map(Dot.Ref(_) --> Dot.Ref(_))*)
 
 extension (inline stringContext: StringContext)
   transparent inline def ref(inline parts: Any*): Dot.Ref =
