+++ lib/dissonance/src/core/dissonance.Diff.scala
@@ -114,6 +114,23 @@ object Diff:
 
     recur(lines, 1, Nil, 0, 0, 0)
 
+  extension (diff: Diff[Text])
+    def serialize: Stream[Text] = diff.chunks.flatMap:
+      case Chunk(left, right, dels, inss) =>
+        def range(start: Int, end: Int): Text =
+          s"$start${if start == end then "" else s",$end"}".tt
+
+        val command: Text =
+          if dels.isEmpty then s"${left}a${range(right + 1, right + inss.size)}".tt
+          else if inss.isEmpty then s"${range(left + 1, left + dels.size)}d${right}".tt
+          else s"${range(left + 1, left + dels.size)}c${range(right + 1, right + inss.size)}".tt
+
+        val delSeq = dels.map { del => Text("< "+del.value) }
+        val sep = if inss.size > 0 && dels.size > 0 then List(Text("---")) else List()
+        val insSeq = inss.map { ins => Text("> "+ins.value) }
+
+        command :: delSeq ::: sep ::: insSeq
+
 case class Diff[element](edits: Edit[element]*):
   def size: Int = edits.count:
     case Par(_, _, _) => false
