+++ lib/honeycomb/src/core/honeycomb_core.scala
@@ -41,8 +41,6 @@ import contextual.*
 import fulminate.*
 import prepositional.*
 import proscenium.*
-import rudiments.*
-import vacuous.*
 
 private given realm: Realm = realm"hc"
 
@@ -55,28 +53,6 @@ package attributives:
 
 extension (inline context: StringContext)
   transparent inline def h: Interpolation = interpolation[Html](context)
-
-extension (html: Seq[Html])
-  def nodes: IArray[Node] =
-    var count = 0
-
-    for item <- html do item match
-      case fragment: Fragment => count += fragment.nodes.length
-      case _                  => count += 1
-
-    val array = new Array[Node](count)
-
-    var index = 0
-    for item <- html do item match
-      case Fragment(nodes*) => for node <- nodes do
-        array(index) = node
-        index += 1
-
-      case node: Node =>
-        array(index) = node
-        index += 1
-
-    array.immutable(using Unsafe)
 
 package doms.html:
   given whatwg: Whatwg = Whatwg()
