+++ lib/xylophone/src/core/xylophone.Xml.scala
@@ -87,6 +87,29 @@ object Xml extends Tag.Container
 
   def header: Header = Header("1.0", Unset, Unset)
 
+  extension (xml: Seq[Xml])
+    def nodes: IArray[Node] =
+      var count = 0
+
+      for item <- xml do item match
+        case fragment: Fragment => count += fragment.nodes.length
+        case _                  => count += 1
+
+      val array = new Array[Node](count)
+
+      var index = 0
+      for item <- xml do item match
+        case Fragment(nodes*) =>
+          for node <- nodes do
+          array(index) = node
+          index += 1
+
+        case node: Node =>
+          array(index) = node
+          index += 1
+
+      array.immutable(using Unsafe)
+
   inline given interpolator: Xml is Interpolable:
     type Result = Xml
