+++ lib/xylophone/src/core/xylophone_core.scala
@@ -51,27 +51,4 @@ extension [encodable: Encodable in Xml](value: encodable)
 extension (inline context: StringContext)
   transparent inline def x: Interpolation = interpolation[Xml](context)
 
-extension (xml: Seq[Xml])
-  def nodes: IArray[Node] =
-    var count = 0
-
-    for item <- xml do item match
-      case fragment: Fragment => count += fragment.nodes.length
-      case _                  => count += 1
-
-    val array = new Array[Node](count)
-
-    var index = 0
-    for item <- xml do item match
-      case Fragment(nodes*) =>
-        for node <- nodes do
-        array(index) = node
-        index += 1
-
-      case node: Node =>
-        array(index) = node
-        index += 1
-
-    array.immutable(using Unsafe)
-
 private given realm: Realm = realm"xy"
